Optimise endgame piece selection and reduce download overhead while in endgame.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8572 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DownloadWithTiger.cpp
@@ -81,7 +81,7 @@ DWORD CDownloadWithTiger::GetValidationCookie() const
 	return m_nVerifyCookie;
 }
 
-QWORD CDownloadWithTiger::GetVerifyLength(PROTOCOLID nProtocol, int nHash) const
+DWORD CDownloadWithTiger::GetVerifyLength(PROTOCOLID nProtocol, int nHash) const
 {
 	CQuickLock oLock( m_pTigerSection );
